Restrict task SQL
Don't allow the following SQL statements in tasks:
- SET
- PRAGMA
- INSTALL
- LOAD

You can still install extensions and set configuration options via
init sql
